GG51 - I fixed an issue where I was using prisma to query the db slightly too soon

+++ backend/src/services/profile.service.js
@@ -84,11 +84,14 @@ export async function updateProfile(profileName, newData)
 }
 
 //Friend functions
-export async function addFriend(friend, user)
+export async function addFriend(friendName, user)
 {
+	const friend = await prisma.user.findUnique({ where: { name: friendName }})
+	if (!friend)
+		throw "No user found"
 	//We need to check the UserUserRelation from both angles as the friend request (User Relation) can be initiated from both users
-	const userRelation1 = await prisma.userUserRelation.findUnique({ where: { senderId: user, receiverId: friend }})
-	const userRelation2 = await prisma.userUserRelation.findUnique({ where: { senderId: friend, receiverId: user }})
+	const userRelation1 = await prisma.userUserRelation.findUnique({ where: { senderId: user, receiverId: friend.id }})
+	const userRelation2 = await prisma.userUserRelation.findUnique({ where: { senderId: friend.id, receiverId: user }})
 	if (userRelation1 || userRelation2) //this means the relation is already in PENDING or FRIENDS state
 		throw "User relation already exists" //error code?
 	await prisma.user.update({
@@ -97,7 +100,7 @@ export async function addFriend(friend, user)
 	sentRequests: {
 		create: [
 			{
-				receiverId: { connect: { id: friend }},
+				receiverId: { connect: { id: friend.id }},
 				status: "PENDING"
 			}
 		]
@@ -106,36 +109,45 @@ export async function addFriend(friend, user)
 	});
 }
 
-export async function acceptFriendRequest(friend, user)
+export async function acceptFriendRequest(friendName, user)
 {
+	const friend = await prisma.user.findUnique({ where: { name: friendName }})
+	if (!friend)
+		throw "No user found"
 	//here we are only interested in userrelation where the other one initiated the relation to the user
-	const userRelation = await prisma.userUserRelation.findUnique({ where: { senderId: friend, receiverId: user}})
+	const userRelation = await prisma.userUserRelation.findUnique({ where: { senderId: friend.id, receiverId: user}})
 	if (!userRelation || userRelation.status != "PENDING")
 		throw "No pending user relation"		
 	await prisma.userUserRelation.update({
-	where: { senderId: friend, receiverId: user}, //now since i am accepting, the friend here is the user
+	where: { senderId: friend.id, receiverId: user}, //now since i am accepting, the friend here is the user
 	data: {
 			status: "FRIENDS"
 	},
 	});
 }
 
-export async function declineFriendRequest(friend, user)
+export async function declineFriendRequest(friendName, user)
 {
-	const userRelation = await prisma.userUserRelation.findUnique({ where: { senderId: friend, receiverId: user}})
+	const friend = await prisma.user.findUnique({ where: { name: friendName }})
+	if (!friend)
+		throw "No user found"
+	const userRelation = await prisma.userUserRelation.findUnique({ where: { senderId: friend.id, receiverId: user}})
 	if (!userRelation || userRelation.status != "PENDING")
 		throw "No pending user relation"		
 	await prisma.userUserRelation.delete({ //As we remove friend, we can remove the entire UserUserRelation db entry
-	where: { senderId: friend, receiverId: user},
+	where: { senderId: friend.id, receiverId: user},
 	});
 }
 
 //remove friend can be done by the initiater to both pending and friends. the accepter can only remove FRIENDS because they have the decline
-export async function removeFriend(friend, user)
+export async function removeFriend(friendName, user)
 {
+	const friend = await prisma.user.findUnique({ where: { name: friendName }})
+	if (!friend)
+		throw "No user found"
 	//We need to check the UserUserRelation from both angles as the friend request (User Relation) can be initiated from both users
-	const userRelation1 = await prisma.userUserRelation.findUnique({ where: { senderId: user, receiverId: friend }})
-	const userRelation2 = await prisma.userUserRelation.findUnique({ where: { senderId: friend, receiverId: user }})
+	const userRelation1 = await prisma.userUserRelation.findUnique({ where: { senderId: user, receiverId: friend.id }})
+	const userRelation2 = await prisma.userUserRelation.findUnique({ where: { senderId: friend.id, receiverId: user }})
 	if (!userRelation1 && !userRelation2)
 	{
 		throw "User relation does not exist"
@@ -147,13 +159,101 @@ export async function removeFriend(friend, user)
 	if (userRelation1)
 	{
 		await prisma.userUserRelation.delete({ //As we remove friend, we can remove the entire UserUserRelation db entry
-		where: { senderId: user, receiverId: friend},
+		where: { senderId: user, receiverId: friend.id},
 	});
 	}
 	if (userRelation2)
 	{
 		await prisma.userUserRelation.delete({ //As we remove friend, we can remove the entire UserUserRelation db entry
-		where: { senderId: friend, receiverId: user},
+		where: { senderId: friend.id, receiverId: user},
 	});
 	}
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+// //Friend functions
+// export async function addFriend(friend, user)
+// {
+// 	//We need to check the UserUserRelation from both angles as the friend request (User Relation) can be initiated from both users
+// 	const userRelation1 = await prisma.userUserRelation.findUnique({ where: { senderId: user, receiverId: friend }})
+// 	const userRelation2 = await prisma.userUserRelation.findUnique({ where: { senderId: friend, receiverId: user }})
+// 	if (userRelation1 || userRelation2) //this means the relation is already in PENDING or FRIENDS state
+// 		throw "User relation already exists" //error code?
+// 	await prisma.user.update({
+// 	where: { id: user },
+// 	data: {
+// 	sentRequests: {
+// 		create: [
+// 			{
+// 				receiverId: { connect: { id: friend }},
+// 				status: "PENDING"
+// 			}
+// 		]
+// 	}
+// 	},
+// 	});
+// }
+
+// export async function acceptFriendRequest(friend, user)
+// {
+// 	//here we are only interested in userrelation where the other one initiated the relation to the user
+// 	const userRelation = await prisma.userUserRelation.findUnique({ where: { senderId: friend, receiverId: user}})
+// 	if (!userRelation || userRelation.status != "PENDING")
+// 		throw "No pending user relation"		
+// 	await prisma.userUserRelation.update({
+// 	where: { senderId: friend, receiverId: user}, //now since i am accepting, the friend here is the user
+// 	data: {
+// 			status: "FRIENDS"
+// 	},
+// 	});
+// }
+
+// export async function declineFriendRequest(friend, user)
+// {
+// 	const userRelation = await prisma.userUserRelation.findUnique({ where: { senderId: friend, receiverId: user}})
+// 	if (!userRelation || userRelation.status != "PENDING")
+// 		throw "No pending user relation"		
+// 	await prisma.userUserRelation.delete({ //As we remove friend, we can remove the entire UserUserRelation db entry
+// 	where: { senderId: friend, receiverId: user},
+// 	});
+// }
+
+// //remove friend can be done by the initiater to both pending and friends. the accepter can only remove FRIENDS because they have the decline
+// export async function removeFriend(friend, user)
+// {
+// 	//We need to check the UserUserRelation from both angles as the friend request (User Relation) can be initiated from both users
+// 	const userRelation1 = await prisma.userUserRelation.findUnique({ where: { senderId: user, receiverId: friend }})
+// 	const userRelation2 = await prisma.userUserRelation.findUnique({ where: { senderId: friend, receiverId: user }})
+// 	if (!userRelation1 && !userRelation2)
+// 	{
+// 		throw "User relation does not exist"
+// 	}
+// 	if (userRelation2 && userRelation2.status == "PENDING") //if you have received a friend request and its pending, you should not be able to remove friend
+// 	{
+// 		throw "No remove action is allowed"
+// 	}
+// 	if (userRelation1)
+// 	{
+// 		await prisma.userUserRelation.delete({ //As we remove friend, we can remove the entire UserUserRelation db entry
+// 		where: { senderId: user, receiverId: friend},
+// 	});
+// 	}
+// 	if (userRelation2)
+// 	{
+// 		await prisma.userUserRelation.delete({ //As we remove friend, we can remove the entire UserUserRelation db entry
+// 		where: { senderId: friend, receiverId: user},
+// 	});
+// 	}
+// }
